% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{augment_data_plot}
\alias{augment_data_plot}
\title{Augment the data for better ploting}
\usage{
augment_data_plot(data_ranked)
}
\arguments{
\item{data_ranked}{the dataset built with \code{\link{compare_last_2_rankings}}}
}
\value{
the dataset with everything needed for the plot (and some more)
}
\description{
Augment the data for better ploting
}
