% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{compare_last_2_rankings}
\alias{compare_last_2_rankings}
\title{Compare ranking between the extra mortality between the last report and the previous one}
\usage{
compare_last_2_rankings(data_ranked)
}
\arguments{
\item{data_ranked}{the dataset built with \code{\link{prepare_data_with_rank}}}
}
\value{
a \code{tibble} with all the relevant data and the ranks for 2 reporting days
}
\description{
Compare ranking between the extra mortality between the last report and the previous one
}
