% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronaR-package.R
\docType{package}
\name{coronaR-package}
\alias{coronaR-package}
\alias{coronaR}
\title{Welcome to coronaR}
\description{
The goal of this package is to make it easy to explore and plot mortality caused by COVID19.
}
\examples{
## See the online README file for examples.
}
\keyword{package}
