% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_deaths}
\alias{plot_deaths}
\title{Create the data for plotting mortality and draw the plot}
\usage{
plot_deaths(
  data_ECDC,
  data_WB,
  type_major = c("daily", "cumul"),
  type_minor = NULL,
  baseline_major = c("country", "world"),
  baseline_minor = NULL,
  select_major = c("worst_day", "last_day"),
  select_minor = NULL,
  alpha_major = 0.5,
  alpha_minor = 1,
  title = "",
  return_plot = FALSE
)
}
\arguments{
\item{data_ECDC}{the dataset built with \code{\link{prepare_data_ECDC}}}

\item{data_WB}{the dataset built with \code{\link{prepare_data_WB}}}

\item{type_major}{either \code{"daily"} (for daily deaths) or \code{"cumul"} (for cumulative deaths)}

\item{type_minor}{same for minor dataset (see Details)}

\item{baseline_major}{either \code{"country"} (for correcting by the baseline mortality in the country), or \code{"world"} (for correcting all countries by the same baseline mortality)}

\item{baseline_minor}{same for minor dataset (see Details)}

\item{select_major}{either \code{"worst_day"} (for measuring mortality at its worst) or \code{"last_day"} (for measuring mortality at the last day in the report considered)}

\item{select_minor}{same for minor dataset (see Details)}

\item{alpha_major}{a number betweeen 0 and 1 defining the transparency}

\item{alpha_minor}{same for minor dataset (see Details)}

\item{title}{the title for the plot}

\item{return_plot}{either FALSE (default) for returning data or TRUE for returning the plot so to modify it}
}
\value{
the data used for the plot (invisibly) or the plot (invisibly) depending on the argument \code{return_plot}
}
\description{
This function calls \code{\link{prepare_data_plot}} (up to) twice to prepare (up to) 2 different
datasets to be plotted.
}
\details{
The so-called major dataset is the one used to rank the country.
The so-called is used to add information within each horizontal bars.
The typical usage is to use the major dataset to represent the mortality at their worst,
and to use the minor dataset to represent the mortality of the last date. Yet,
this can be used to represent other combinations.
}
