% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_data_plot}
\alias{prepare_data_plot}
\title{A wrapper to faciliate the preparation of the data for the plot}
\usage{
prepare_data_plot(
  data_ECDC,
  data_WB,
  type = c("daily", "cumul"),
  baseline = c("country", "world"),
  select = c("worst_day", "last_day")
)
}
\arguments{
\item{data_ECDC}{the dataset built with \code{\link{prepare_data_ECDC}}}

\item{data_WB}{the dataset built with \code{\link{prepare_data_WB}}}

\item{type}{either \code{"daily"} (for daily deaths) or \code{"cumul"} (for cumulative deaths)}

\item{baseline}{either \code{"country"} (for correcting by the baseline mortality in the country), or \code{"world"} (for correcting all countries by the same baseline mortality)}

\item{select}{either \code{"worst_day"} (for measuring mortality at its worst) or \code{"last_day"} (for measuring mortality at the last day in the report considered)}
}
\value{
the dataset with just what is needed for the plot (and no more)
}
\description{
This function is a wrapper successively calling \code{\link{merge_datasets}},
\code{\link{prepare_data_with_rank}}, \code{\link{compare_last_2_rankings}}, \code{\link{augment_data_plot}}
and then selecting what is needed for the plot.
}
