% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_WB.R
\name{prepare_data_WB}
\alias{prepare_data_WB}
\title{Download and prepare the data from World Bank}
\usage{
prepare_data_WB(min_pop_size = 2e+06)
}
\arguments{
\item{min_pop_size}{population size below which data will be dropped}
}
\value{
a tidy WB dataset
}
\description{
Download and prepare the data from World Bank
}
