% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{prepare_data_with_rank}
\alias{prepare_data_with_rank}
\title{Compute the ranks based on extra mortality and simplify data}
\usage{
prepare_data_with_rank(data_combined)
}
\arguments{
\item{data_combined}{the dataset built with \code{\link{merge_datasets}}}
}
\value{
a \code{tibble} with all the relevant data and the ranks for all days
}
\description{
Compute the ranks based on extra mortality and simplify data
}
\note{
What the extra mortality represents depends on what has been done with \code{\link{merge_datasets}}
}
