% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{safe_cummax}
\alias{safe_cummax}
\title{Compute the cumulative maximum without failing}
\usage{
safe_cummax(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a numerical vector
}
\description{
This function attempts to solve some limitations of \code{\link{cummax}} as shown in example.
}
\examples{
safe_cummax(x = c(NA, NA, 1, 1, 2, 2, 3))
safe_cummax(x = c(NA, NA))

}
