% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{safe_max}
\alias{safe_max}
\title{Compute the maximum without failing}
\usage{
safe_max(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a scalar
}
\description{
This function attempts to solve some limitations of \code{\link{max}} as shown in example.
}
\examples{
safe_max(c(NA, NA))
safe_max(c(1, 2))
safe_max(c(NA, 1, 2))

}
